+++ examples/nf_router/go.sh
@@ -2,19 +2,35 @@
 
 function usage {
         echo "$0 CPU-LIST SERVICE-ID ROUTER_CONF [-p PRINT] [-n NF-ID]"
+        echo "$0 -F CONFIG_FILE -- -- -f ROUTER_CONF [-p PRINT]"
+        echo ""
         echo "$0 3 1 router.conf --> cores 3 with Service ID 1, and router config 'router.conf'"
+        echo "$0 -F example_config.json -- -- -f router.conf --> load from example_config.json and use router.conf"
         exit 1
 }
 
 SCRIPT=$(readlink -f "$0")
 SCRIPTPATH=$(dirname "$SCRIPT")
+
+if [ -z $1 ]
+then
+  usage
+fi
+
+if [ $1 = "-F" ]
+then
+  config=$2
+  shift 2
+  exec sudo $SCRIPTPATH/build/app/nf_router -F $config "$@"
+fi
+
 cpu=$1
 service=$2
-config=$3
+router_config=$3
 
 shift 3
 
-if [ -z $config ]
+if [ -z $router_config ]
 then
     usage
 fi
@@ -28,4 +44,4 @@ while getopts ":p:n:" opt; do
   esac
 done
 
-exec sudo $SCRIPTPATH/build/app/nf_router -l $cpu -n 3 --proc-type=secondary -- -r $service $instance -- -f $config $print
+exec sudo $SCRIPTPATH/build/app/nf_router -l $cpu -n 3 --proc-type=secondary -- -r $service $instance -- -f $router_config $print